requirements.txt
gradio==4.15.0
openai-whisper
ffmpeg-python
torch
python-docx
pandas
